requirements.txt
flask==2.0.1
bcrypt==3.2.0
pyjwt==2.1.0
scapy==2.4.5
python-nmap==0.7.1
netifaces==0.11.0
psutil==5.8.0
requests==2.26.0
cryptography==3.4.7
torch==1.9.0
transformers==4.9.2
scikit-learn==0.24.2
gensim==4.0.1
librosa==0.8.1
matplotlib==3.4.3
networkx==2.6.2
